Create useAuth hook to improve reusability of sign-in and sign-up components (Finish)
