+++ client/src/features/ui/form-text-field/form-text-field.component.tsx
@@ -1,23 +1,27 @@
-import { FC, RefObject } from "react";
+import { FC, RefObject, HTMLInputTypeAttribute } from "react";
 import { TextField } from "@mui/material";
 
 interface IFormTextField {
   errorMessage?: string;
+  errorType: "username" | "email" | "password";
+  type: HTMLInputTypeAttribute;
   ref?: RefObject<HTMLInputElement>;
-  type: string;
   label: string;
 }
 
 export const FormTextField: FC<IFormTextField> = ({
   errorMessage,
-  ref,
+  errorType,
   type,
+  ref,
   label,
 }) => {
   return (
     <TextField
-      error={errorMessage?.toLowerCase().includes("email")}
-      helperText={errorMessage?.toLowerCase().includes("email") && errorMessage}
+      error={errorMessage?.toLowerCase().includes(errorType)}
+      helperText={
+        errorMessage?.toLowerCase().includes(errorType) && errorMessage
+      }
       inputRef={ref}
       label={label}
       variant="standard"
