+++ client/src/pages/sign-up.tsx
@@ -1,88 +1,66 @@
-import { FC, useRef, FormEvent, useState } from "react";
+import { FC } from "react";
 import "./sign-up.style.css";
-import { Link, Redirect, useHistory } from "react-router-dom";
-import { Button, Box, TextField, CircularProgress } from "@mui/material";
-import { useAppDispatch, useAppSelector } from "@redux/hooks";
-import { selectUser, signUpUser } from "@redux/user.slice";
-import { openAlert } from "@redux/alert.slice";
+import { Link, Redirect } from "react-router-dom";
+import { Button, Box, CircularProgress } from "@mui/material";
+import { useAppSelector } from "@redux/hooks";
+import { selectUser } from "@redux/user.slice";
 import { selectFriends } from "@redux/friends.slice";
-import { MainLoading } from "@features/ui";
+import { FormTextField, MainLoading } from "@features/ui";
 import { useRedirect } from "@hooks/use-redirect";
+import { useAuth } from "@hooks/use-auth";
+
+interface ISignUpButton {
+  isSignUp: boolean;
+}
+
+const SignUpAndBackToSignInButtons: FC<ISignUpButton> = ({ isSignUp }) => {
+  return (
+    <>
+      <Button
+        type="submit"
+        sx={{
+          width: "100%",
+          borderRadius: "6px",
+        }}
+        variant="contained"
+        startIcon={
+          isSignUp && (
+            <CircularProgress size={20} sx={{ color: "white", my: 0.282 }} />
+          )
+        }
+      >
+        {!isSignUp && <span>SIGN UP</span>}
+      </Button>
+      <Box sx={{ height: 15 }} />
+      <Link to="/" style={{ textDecoration: "none" }}>
+        <Button
+          sx={{
+            width: "100%",
+            borderRadius: "6px",
+          }}
+          variant="contained"
+          className="signup__back"
+        >
+          BACK
+        </Button>
+      </Link>
+    </>
+  );
+};
 
 export const SignUp: FC = () => {
-  const emailRef = useRef<HTMLInputElement>(null);
-  const usernameRef = useRef<HTMLInputElement>(null);
-  const passwordRef = useRef<HTMLInputElement>(null);
-  const passwordConfirmRef = useRef<HTMLInputElement>(null);
-
-  const [errorMessage, setErrorMessage] = useState("");
-
-  const history = useHistory();
-
-  const dispatch = useAppDispatch();
-
   const user = useAppSelector(selectUser);
   const friends = useAppSelector(selectFriends);
 
   const { pathname } = useRedirect();
-
-  const signUpHandler = async (event: FormEvent<HTMLFormElement>) => {
-    event.preventDefault();
-
-    const email = emailRef.current!.value;
-    const userName = usernameRef.current!.value;
-    const password = passwordRef.current!.value;
-    const passwordConfirm = passwordConfirmRef.current!.value;
-
-    if (!email) {
-      setErrorMessage("Please fill in email");
-      return;
-    }
-
-    if (!userName) {
-      setErrorMessage("Please fill in username");
-      return;
-    }
-
-    if (!password || !passwordConfirm) {
-      setErrorMessage("Please fill in passwords");
-      return;
-    }
-
-    if (password !== passwordConfirm) {
-      setErrorMessage("Passwords do not match");
-      return;
-    }
-
-    if (password === passwordConfirm && password.length < 6) {
-      setErrorMessage("Passwords must be at least 6 characters");
-      return;
-    }
-
-    setErrorMessage("");
-
-    try {
-      const dispatchResult = await dispatch(
-        signUpUser({
-          email,
-          userName,
-          password,
-        })
-      ).unwrap();
-
-      if (dispatchResult.code === "error") {
-        setErrorMessage(dispatchResult.message);
-      } else if (dispatchResult.code === "success") {
-        dispatch(openAlert({ message: "Sign up successful!" }));
-
-        setTimeout(() => {
-          history.push("/"); // redirect back to sign in page
-        }, 2000);
-      }
-    } catch (error) {
-      console.error(error);
-    }
-  };
+  const {
+    signUpHandler,
+    errorMessage,
+    emailRef,
+    usernameRef,
+    passwordRef,
+    passwordConfirmRef,
+  } = useAuth("sign-up");
 
   return (
     <>
@@ -99,91 +77,39 @@ export const SignUp: FC = () => {
             <h1 className="signup__heading">SIGN UP</h1>
             <div className="signup__field">
               <form className="signup__form" onSubmit={signUpHandler}>
-                {/* Inputs */}
-                <TextField
-                  error={errorMessage.toLowerCase().includes("email")}
-                  helperText={
-                    errorMessage.toLowerCase().includes("email") && errorMessage
-                  }
-                  inputRef={emailRef}
+                <FormTextField
+                  errorMessage={errorMessage}
+                  errorType="email"
+                  ref={emailRef}
                   label="Email"
-                  variant="standard"
-                  sx={{ width: "100%" }}
                   type="email"
                 />
                 <Box sx={{ height: 20 }} />
-                <TextField
-                  error={errorMessage.toLowerCase().includes("username")}
-                  helperText={
-                    errorMessage.toLowerCase().includes("username") &&
-                    errorMessage
-                  }
-                  inputRef={usernameRef}
-                  type="text"
+                <FormTextField
+                  errorMessage={errorMessage}
+                  errorType="username"
+                  ref={usernameRef}
                   label="Username"
-                  variant="standard"
-                  sx={{ width: "100%" }}
+                  type="text"
                 />
                 <Box sx={{ height: 20 }} />
-                <TextField
-                  error={errorMessage.toLowerCase().includes("password")}
-                  helperText={
-                    errorMessage.toLowerCase().includes("password") &&
-                    errorMessage
-                  }
-                  inputRef={passwordRef}
-                  type="password"
+                <FormTextField
+                  errorMessage={errorMessage}
+                  errorType="password"
+                  ref={passwordRef}
                   label="Password"
-                  variant="standard"
-                  sx={{ width: "100%" }}
+                  type="password"
                 />
                 <Box sx={{ height: 20 }} />
-                <TextField
-                  error={errorMessage.toLowerCase().includes("password")}
-                  helperText={
-                    errorMessage.toLowerCase().includes("password") &&
-                    errorMessage
-                  }
-                  inputRef={passwordConfirmRef}
-                  type="password"
+                <FormTextField
+                  errorMessage={errorMessage}
+                  errorType="password"
+                  ref={passwordConfirmRef}
                   label="Confirm password"
-                  variant="standard"
-                  sx={{ width: "100%" }}
+                  type="password"
                 />
                 <Box sx={{ height: 40 }} />
-
-                {/* Buttons */}
-                <Button
-                  type="submit"
-                  sx={{
-                    width: "100%",
-                    borderRadius: "6px",
-                  }}
-                  variant="contained"
-                  startIcon={
-                    user.isSignUp && (
-                      <CircularProgress
-                        size={20}
-                        sx={{ color: "white", my: 0.282 }}
-                      />
-                    )
-                  }
-                >
-                  {!user.isSignUp && <span>SIGN UP</span>}
-                </Button>
-                <Box sx={{ height: 15 }} />
-                <Link to="/" style={{ textDecoration: "none" }}>
-                  <Button
-                    sx={{
-                      width: "100%",
-                      borderRadius: "6px",
-                    }}
-                    variant="contained"
-                    className="signup__back"
-                  >
-                    BACK
-                  </Button>
-                </Link>
+                <SignUpAndBackToSignInButtons isSignUp={user.isSignUp} />
               </form>
             </div>
           </div>
